Refactor: DTO 패키지 구조를 request / response 분리 및 재구성
