+++ src/main/java/com/codeit/mission/deokhugam/book/dto/request/BookCreateRequest.java
@@ -1,8 +1,8 @@
-package com.codeit.mission.deokhugam.book.dto;
+package com.codeit.mission.deokhugam.book.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import java.time.LocalDate;
+import java.time.Instant;
 
 public record BookCreateRequest(
     //제목
@@ -23,7 +23,7 @@ public record BookCreateRequest(
 
     //출판일
     @NotNull(message = "출판 일자를 적어주세요.")
-    LocalDate publishedDate,
+    Instant publishedDate,
 
     //ISBN 값
     String isbn
