+++ src/main/java/com/codeit/mission/deokhugam/book/dto/response/CursorPageResponseBookDto.java
@@ -1,12 +1,12 @@
-package com.codeit.mission.deokhugam.book.dto;
+package com.codeit.mission.deokhugam.book.dto.response;
 
-import java.time.LocalDateTime;
+import java.time.Instant;
 import java.util.List;
 
 public record CursorPageResponseBookDto(
     List<BookDto> content,
     String nextCursor,
-    LocalDateTime nextAfter,
+    Instant nextAfter,
     int size,
     Long totalElements,
     boolean hasNext
