+++ src/main/java/com/codeit/mission/deokhugam/book/dto/response/NaverResponseDto.java
@@ -1,4 +1,4 @@
-package com.codeit.mission.deokhugam.book.dto;
+package com.codeit.mission.deokhugam.book.dto.response;
 
 import java.util.List;
