+++ src/main/java/com/codeit/mission/deokhugam/book/dto/response/OcrResponse.java
@@ -1,4 +1,4 @@
-package com.codeit.mission.deokhugam.book.dto;
+package com.codeit.mission.deokhugam.book.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
